session for Admin & user ok on manageArticle
securizing url  with auth session ok

+++ App/Controller/ArticleController.php
@@ -27,11 +27,17 @@ class ArticleController
     {
         $twig = new TwigHelper();
         $functionHelper = new FunctionHelper();
-        $title = "Publier un article";
 
-        $user = $functionHelper->checkActiveUserInSession();
+        $sessionOK = $functionHelper->mustBeAuthentificated();
+        var_dump($sessionOK);
+        if ($sessionOK) {
+            $title = "Publier un article";
 
-        $twig->loadTwig()->display('article/formAddArticle.html.twig', ['message' => $message, 'user' => $user, 'title' => $title]);
+            $user = $functionHelper->checkActiveUserInSession();
+
+            $twig->loadTwig()->display('article/formAddArticle.html.twig', ['message' => $message, 'user' => $user, 'title' => $title]);
+        }
+
     }
 
     public static function showFormUpdateArticle($slug)
@@ -40,31 +46,43 @@ class ArticleController
         $functionHelper = new FunctionHelper();
         $articleManager = new ArticleManager();
 
-        $user = $functionHelper->checkActiveUserInSession();
+        $sessionOK = $functionHelper->mustBeAuthentificated();
 
-        if (is_array($slug)) {
-            $article = $articleManager->selectOneArticle($slug['slug']);
-            $message =  ['error' => $slug['error']];
-        } else {
-            $article = $articleManager->selectOneArticle($slug);
-            $message = null;
+        if ($sessionOK) {
+            $user = $functionHelper->checkActiveUserInSession();
+
+            if (is_array($slug)) {
+                $article = $articleManager->selectOneArticle($slug['slug']);
+                $message = ['error' => $slug['error']];
+            } else {
+                $article = $articleManager->selectOneArticle($slug);
+                $message = null;
+            }
+            $articleTitle = $article["title"];
+            $title = "Mettre à jour l'article : $articleTitle";
+
+            $twig->loadTwig()->display('article/formUpdateArticle.html.twig', ['message' => $message, 'user' => $user, 'article' => $article, 'title' => $title]);
         }
-        $articleTitle = $article["title"];
-        $title = "Mettre à jour l'article : $articleTitle";
-
-        $twig->loadTwig()->display('article/formUpdateArticle.html.twig', ['message' => $message, 'user' => $user, 'article' => $article, 'title' => $title]);
     }
-
 
     public static function manageArticles()
     {
         $twig = new TwigHelper();
         $functionHelper = new FunctionHelper();
         $articleManager = new ArticleManager();
+        $role = $_SESSION['userRole'];
+        $sessionOK = $functionHelper->mustBeAuthentificated();
 
-        $user = $functionHelper->checkActiveUserInSession();
-        $articles = $articleManager->selectArticleByUser($user);
-        $twig->loadTwig()->display('article/manageArticles.html.twig', ['user' => $user, 'articles' => $articles]);
+        if ($sessionOK) {
+
+            $user = $functionHelper->checkActiveUserInSession();
+            if ($role === 'admin') {
+                $articles = $articleManager->selectAllArticles();
+            } else {
+                $articles = $articleManager->selectArticleByUser($user);
+            }
+            $twig->loadTwig()->display('article/manageArticles.html.twig', ['user' => $user, 'articles' => $articles]);
+        }
     }
 
     public static function addArticle(array $data = [])
@@ -73,57 +91,60 @@ class ArticleController
         $request = new  Request();
         $pathUploadDir = '../public/images/articles/';
         $uniq = uniqid();
+        $sessionOK = $functionHelper->mustBeAuthentificated();
 
         try {
-            $newDirPath = "$pathUploadDir$uniq";
-            $title = $data['title'];
-            $tags = trim($data['tag']);
-            $content = $data['content'];
+            if ($sessionOK) {
+                $newDirPath = "$pathUploadDir$uniq";
+                $title = $data['title'];
+                $tags = trim($data['tag']);
+                $content = $data['content'];
 
-            if (!empty($tags) && strpos($tags, ';') === false) {
-                $request->redirectToRoute('newPost', ['error' => 'Veuillez remplir le champ tag comme suivi => bateau;chat;chocolat']);
-            } elseif (strlen($title) > 255 || strlen($tags) > 255 || strlen($_FILES['image']['name']) > 255) {
-                $request->redirectToRoute('register', ['error' => "Le champ et le nom de l'image doit être inférieur à 255 caractères"]);
-            } else {
-                if (!empty($title) && !empty($content)) {
-
-                    $slug = strtolower(preg_replace('/\s+/', '-', $title));
-                    $slugReady = $functionHelper->removeSpecialAndAccent($slug);
-                    $titleReady = $functionHelper->avoidSqlErrorForString($title);
-                    $contentReady = $functionHelper->avoidSqlErrorForString($content);
-
-                    if ($_FILES['image']['size'] != 0) {
-                        $slugImageToSlug = $functionHelper->uploadImage($newDirPath);
-                    } else {
-                        $slugImageToSlug = false;
-                    }
-                    if ($slugImageToSlug === false) {
-                        $request->redirectToRoute('newPost', ['error' => "L'ajout d'image est obligatoire et doit au être format JPG"]);
-                    } else {
-                        $datePublished = new \DateTime('NOW');
-                        $datePublished = $datePublished->setTimezone(new \DateTimeZone('Europe/Paris'));
-                        $author = $_SESSION['userId'];
-
-                        $articleManager = new ArticleManager();
-                        $registredTitle = $articleManager->selectOneArticleByTitle($titleReady);
-
-                        if ($registredTitle) {
-                            $request->redirectToRoute('newPost', ['error' => "Le titre : $title est déjà utilisé"]);
-                        } else {
-                            $articleManager->insertArticle(
-                                $titleReady,
-                                $slugReady,
-                                $tags,
-                                implode($slugImageToSlug),
-                                $contentReady,
-                                $datePublished->format('Y-m-d H:i:sP'),
-                                $author
-                            );
-                            $request->redirectToRoute('blogIndex', ['success' => "L'article : '$title' a été publié !"]);
-                        }
-                    }
+                if (!empty($tags) && strpos($tags, ';') === false) {
+                    $request->redirectToRoute('newPost', ['error' => 'Veuillez remplir le champ tag comme suivi => bateau;chat;chocolat']);
+                } elseif (strlen($title) > 255 || strlen($tags) > 255 || strlen($_FILES['image']['name']) > 255) {
+                    $request->redirectToRoute('register', ['error' => "Le champ et le nom de l'image doit être inférieur à 255 caractères"]);
                 } else {
-                    $request->redirectToRoute('newPost', ['error' => 'Un article doit comporter obligatoirement un titre et un contenu']);
+                    if (!empty($title) && !empty($content)) {
+
+                        $slug = strtolower(preg_replace('/\s+/', '-', $title));
+                        $slugReady = $functionHelper->removeSpecialAndAccent($slug);
+                        $titleReady = $functionHelper->avoidSqlErrorForString($title);
+                        $contentReady = $functionHelper->avoidSqlErrorForString($content);
+
+                        if ($_FILES['image']['size'] != 0) {
+                            $slugImageToSlug = $functionHelper->uploadImage($newDirPath);
+                        } else {
+                            $slugImageToSlug = false;
+                        }
+                        if ($slugImageToSlug === false) {
+                            $request->redirectToRoute('newPost', ['error' => "L'ajout d'image est obligatoire et doit au être format JPG"]);
+                        } else {
+                            $datePublished = new \DateTime('NOW');
+                            $datePublished = $datePublished->setTimezone(new \DateTimeZone('Europe/Paris'));
+                            $author = $_SESSION['userId'];
+
+                            $articleManager = new ArticleManager();
+                            $registredTitle = $articleManager->selectOneArticleByTitle($titleReady);
+
+                            if ($registredTitle) {
+                                $request->redirectToRoute('newPost', ['error' => "Le titre : $title est déjà utilisé"]);
+                            } else {
+                                $articleManager->insertArticle(
+                                    $titleReady,
+                                    $slugReady,
+                                    $tags,
+                                    implode($slugImageToSlug),
+                                    $contentReady,
+                                    $datePublished->format('Y-m-d H:i:sP'),
+                                    $author
+                                );
+                                $request->redirectToRoute('blogIndex', ['success' => "L'article : '$title' a été publié !"]);
+                            }
+                        }
+                    } else {
+                        $request->redirectToRoute('newPost', ['error' => 'Un article doit comporter obligatoirement un titre et un contenu']);
+                    }
                 }
             }
         } catch
@@ -136,25 +157,28 @@ class ArticleController
     {
         $articleManager = new ArticleManager();
         $request = new Request();
+        $functionHelper = new FunctionHelper();
         $pathUploadDir = '../public/images/';
+        $sessionOK = $functionHelper->mustBeAuthentificated();
 
         try {
+            if ($sessionOK) {
+                $imageArray = $articleManager->selectImagePath($slug);
 
-            $imageArray = $articleManager->selectImagePath($slug);
-
-            if ($imageArray != null) {
-                $imagePath = $imageArray['image'];
-                $pathFile = "$pathUploadDir$imagePath";
-                if (file_exists($pathFile)) {
-                    $folder = substr($imagePath, 0, strpos($imagePath, '/', 10));
-                    $folderPath = "$pathUploadDir$folder";
-                    unlink($pathFile);
-                    rmdir($folderPath);
+                if ($imageArray != null) {
+                    $imagePath = $imageArray['image'];
+                    $pathFile = "$pathUploadDir$imagePath";
+                    if (file_exists($pathFile)) {
+                        $folder = substr($imagePath, 0, strpos($imagePath, '/', 10));
+                        $folderPath = "$pathUploadDir$folder";
+                        unlink($pathFile);
+                        rmdir($folderPath);
+                    }
                 }
-            }
 
-            $articleManager->deleteArticle($slug);
-            $request->redirectToRoute('blogIndex', ['success' => "L'article a bien été supprimé"]);
+                $articleManager->deleteArticle($slug);
+                $request->redirectToRoute('blogIndex', ['success' => "L'article a bien été supprimé"]);
+            }
         } catch (Exception $e) {
             $request->redirectToRoute('blogIndex', ['error' => "Erreur lors de la suppression de l'article $e"]);
         }
@@ -169,83 +193,87 @@ class ArticleController
         $functionHelper = new FunctionHelper;
         $request = new  Request();
         $pathUploadDir = "../public/images/";
+        $sessionOK = $functionHelper->mustBeAuthentificated();
 
 
         try {
-            $title = $data['title'];
-            $tags = trim($data['tag']);
-            $content = $data['content'];
-            if (!empty($tags) && strpos($tags, ';') === false) {
-                $request->redirectToRoute('showFormUpdateArticle',
-                    [
-                        'slug' => $slug,
-                        'error' => 'Veuillez remplir le champ tag comme suivi => bateau;chat;chocolat',
-                    ]);
-            } else {
-                if (strlen($title) > 255 || strlen($tags) > 255 || strlen($_FILES['image']['name']) > 255) {
+            if ($sessionOK) {
+
+                $title = $data['title'];
+                $tags = trim($data['tag']);
+                $content = $data['content'];
+
+
+                if (!empty($tags) && strpos($tags, ';') === false) {
                     $request->redirectToRoute('showFormUpdateArticle',
                         [
                             'slug' => $slug,
-                            'error' => "Le champ et le nom de l'image doit être inférieur à 255 caractères",
+                            'error' => 'Veuillez remplir le champ tag comme suivi => bateau;chat;chocolat',
                         ]);
                 } else {
-                    if (!empty($title) && !empty($content)) {
-                        if (empty($tags)){
-                            $tags = ' ';
-                        }
-                        $slug = strtolower(preg_replace('/\s+/', '-', $title));
-                        $slugReady = $functionHelper->removeSpecialAndAccent($slug);
-                        $titleReady = $functionHelper->avoidSqlErrorForString($title);
-                        $contentReady = $functionHelper->avoidSqlErrorForString($content);
+                    if (strlen($title) > 255 || strlen($tags) > 255 || strlen($_FILES['image']['name']) > 255) {
+                        $request->redirectToRoute('showFormUpdateArticle',
+                            [
+                                'slug' => $slug,
+                                'error' => "Le champ et le nom de l'image doit être inférieur à 255 caractères",
+                            ]);
+                    } else {
+                        if (!empty($title) && !empty($content)) {
+                            if (empty($tags)) {
+                                $tags = ' ';
+                            }
+                            $slug = strtolower(preg_replace('/\s+/', '-', $title));
+                            $slugReady = $functionHelper->removeSpecialAndAccent($slug);
+                            $titleReady = $functionHelper->avoidSqlErrorForString($title);
+                            $contentReady = $functionHelper->avoidSqlErrorForString($content);
 
-                        if ($_FILES['image']['size'] != 0) {
-                            $fileImage = "$pathUploadDir$imagePath";
-                            if (file_exists($fileImage)) {
-                                $folder = substr($imagePath, 0, strpos($imagePath, '/', 10));
-                                $folderPath = "$pathUploadDir$folder";
-                                unlink($fileImage);
-                                $slugImageToSlug = $functionHelper->uploadImage($folderPath);
-                                if ($slugImageToSlug === false) {
+                            if ($_FILES['image']['size'] != 0) {
+                                $fileImage = "$pathUploadDir$imagePath";
+                                if (file_exists($fileImage)) {
+                                    $folder = substr($imagePath, 0, strpos($imagePath, '/', 10));
+                                    $folderPath = "$pathUploadDir$folder";
+                                    unlink($fileImage);
+                                    $slugImageToSlug = $functionHelper->uploadImage($folderPath);
+                                    if ($slugImageToSlug === false) {
+                                        $request->redirectToRoute('showFormUpdateArticle',
+                                            [
+                                                'slug' => $slug,
+                                                'error' => "L'ajout d'image doit être au format JPG",
+                                            ]);
+                                    }
+                                } else {
                                     $request->redirectToRoute('showFormUpdateArticle',
                                         [
                                             'slug' => $slug,
-                                            'error' => "L'ajout d'image doit être au format JPG",
+                                            'error' => "L'article enregistré ne possède pas d'image",
                                         ]);
                                 }
+
                             } else {
+                                $slugImageToSlug[] = $imagePath;
+                            }
+                            $registredTitle = $articleManager->selectOneArticleByTitle($titleReady);
+
+                            if (!empty($registredTitle) && $registredTitle['title'] === $titleReady && $idArticle['id_article'] != $registredTitle['id_article']) {
                                 $request->redirectToRoute('showFormUpdateArticle',
                                     [
                                         'slug' => $slug,
-                                        'error' => "L'article enregistré ne possède pas d'image",
+                                        'error' => "Le titre : $title est déjà utilisé",
                                     ]);
+                            } else {
+                                $articleManager->updateArticle(
+                                    $idArticle,
+                                    $titleReady,
+                                    $slugReady,
+                                    $tags,
+                                    implode($slugImageToSlug),
+                                    $contentReady
+                                );
+                                $request->redirectToRoute('blogIndex', ['success' => "L'article '$titleReady' a bien été mis à jour"]);
                             }
-
-                        } else {
-                            $slugImageToSlug[] = $imagePath;
-                        }
-                        $registredTitle = $articleManager->selectOneArticleByTitle($titleReady);
-
-                        if (!empty($registredTitle) && $registredTitle['title'] === $titleReady && $idArticle['id_article'] != $registredTitle['id_article']) {
-                            $request->redirectToRoute('showFormUpdateArticle',
-                                [
-                                    'slug' => $slug,
-                                    'error' => "Le titre : $title est déjà utilisé",
-                                ]);
-                        } else {
-                            var_dump($tags);
-                            $articleManager->updateArticle(
-                                $idArticle,
-                                $titleReady,
-                                $slugReady,
-                                $tags,
-                                implode($slugImageToSlug),
-                                $contentReady
-                            );
-                            $request->redirectToRoute('blogIndex', ['success' => "L'article '$titleReady' a bien été mis à jour"]);
                         }
                     }
                 }
-
             }
         } catch (Exception $e) {
             $request->redirectToRoute('showFormUpdateArticle', ['error' => "Erreur lors de la mis à jour de l'article $e"]);
